1.6 pre14

+++ src/main/java/icyllis/modernui/ui/master/ModernContainerScreen.java
@@ -82,30 +82,30 @@ public final class ModernContainerScreen<G extends Container> extends ContainerS
 
     @Override
     public final void mouseMoved(double mouseX, double mouseY) {
-        manager.sMouseMoved((int) mouseX, (int) mouseY);
+        manager.sMouseMoved(mouseX, mouseY);
     }
 
     @Override
     public boolean mouseClicked(double mouseX, double mouseY, int mouseButton) {
         super.mouseClicked(mouseX, mouseY, mouseButton);
-        return manager.sMouseClicked((int) mouseX, (int) mouseY, mouseButton);
+        return manager.sMouseClicked(mouseX, mouseY, mouseButton);
     }
 
     @Override
     public boolean mouseReleased(double mouseX, double mouseY, int mouseButton) {
         super.mouseReleased(mouseX, mouseY, mouseButton);
-        return manager.sMouseReleased((int) mouseX, (int) mouseY, mouseButton);
+        return manager.sMouseReleased(mouseX, mouseY, mouseButton);
     }
 
     @Override
     public boolean mouseDragged(double mouseX, double mouseY, int mouseButton, double deltaX, double deltaY) {
         super.mouseDragged(mouseX, mouseY, mouseButton, deltaX, deltaY);
-        return manager.sMouseDragged((int) mouseX, (int) mouseY, deltaX, deltaY);
+        return manager.sMouseDragged(mouseX, mouseY, deltaX, deltaY);
     }
 
     @Override
     public boolean mouseScrolled(double mouseX, double mouseY, double delta) {
-        return manager.sMouseScrolled((int) mouseX, (int) mouseY, delta);
+        return manager.sMouseScrolled(mouseX, mouseY, delta);
     }
 
     @Override
